+++ analyzer/src/load_audit_logs_to_db.js
@@ -1,9 +1,9 @@
 /**
- * @file write transformed_logs.json into sqlite db file audit_log.db
+ * @file write transformed_audit_log_metadata.json into sqlite db file audit_log.db
  */
 const sqlite3 = require("sqlite3").verbose();
 const db = new sqlite3.Database("audit_log.db");
-const transformedLogs = require("../downloads/transformed_logs.json");
+const transformedLogs = require("../downloads/transformed_audit_log_metadata.json");
 
 db.serialize(() => {
   db.run(`CREATE TABLE IF NOT EXISTS audit_log (
